Refactor tests for TPU Accelerator
